@@ -9,13 +9,13 @@
     "node": ">=20.9.0"
   },
   "scripts": {
-    "build": "run-s -l build:angular1 build:angular17",
+    "build": "run-s -l build:angular1 build:angular18",
     "build:angular1": "grunt build",
-    "build:angular17": "ng build",
+    "build:angular18": "ng build",
     "lint:fix": "ng lint --fix",
     "lint": "ng lint",
-    "serve:angular17": "export NODE_OPTIONS=--max_old_space_size=4096 && ng serve --configuration $NODE_ENV",
-    "start": "npm-run-all -l -s build:angular1 -p watch:angular1 serve:angular17",
+    "serve:angular18": "export NODE_OPTIONS=--max_old_space_size=4096 && ng serve --configuration $NODE_ENV",
+    "start": "npm-run-all -l -s build:angular1 -p watch:angular1 serve:angular18",
     "watch:angular1": "grunt delta",
     "deploy:build2api": "ng build --delete-output-path=true --optimization=true --configuration production --output-path dist",
     "deploy": "run-s -l build:angular1 deploy:build2api",
@@ -26,20 +26,19 @@
   "keywords": [],
   "author": "",
   "dependencies": {
-    "@angular/animations": "^17.3.6",
-    "@angular/cdk": "^17.3.6",
-    "@angular/cli": "^17.3.6",
-    "@angular/common": "^17.3.6",
-    "@angular/compiler": "^17.3.6",
-    "@angular/core": "^17.3.6",
-    "@angular/forms": "^17.3.6",
-    "@angular/material": "^17.3.6",
-    "@angular/material-moment-adapter": "^17.3.6",
-    "@angular/platform-browser": "^17.3.6",
-    "@angular/platform-browser-dynamic": "^17.3.6",
-    "@angular/router": "^17.3.6",
-    "@angular/service-worker": "^17.3.6",
-    "@angular/upgrade": "^17.3.6",
+    "@angular/animations": "^18.0.1",
+    "@angular/cdk": "^18.0.1",
+    "@angular/common": "^18.0.1",
+    "@angular/compiler": "^18.0.1",
+    "@angular/core": "^18.0.1",
+    "@angular/forms": "^18.0.1",
+    "@angular/material": "^18.0.1",
+    "@angular/material-moment-adapter": "^18.0.1",
+    "@angular/platform-browser": "^18.0.1",
+    "@angular/platform-browser-dynamic": "^18.0.1",
+    "@angular/router": "^18.0.1",
+    "@angular/service-worker": "^18.0.1",
+    "@angular/upgrade": "^18.0.1",
     "@ctrl/ngx-emoji-mart": "^9.2.0",
     "@ngneat/hotkeys": "^4.0.0",
     "@swimlane/ngx-charts": "^20.5.0",
@@ -80,8 +79,8 @@
     "moment": "^2.29.4",
     "ng-csv": "0.2.3",
     "ng-file-upload": "~5.0.9",
-    "ng-flex-layout": "^17.3.4-beta.1",
-    "ng2-pdf-viewer": "^10.0",
+    "ng-flex-layout": "^17.3.7-beta.1",
+    "ng2-pdf-viewer": "^10.2",
     "ngx-bootstrap": "^6.1.0",
     "ngx-entity-service": "^0.0.39",
     "ngx-lottie": "^11.0.2",
@@ -93,14 +92,10 @@
     "zone.js": "~0.14"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^17.3.6",
-    "@angular-eslint/builder": "^17.3.0",
-    "@angular-eslint/eslint-plugin": "^17.3.0",
-    "@angular-eslint/eslint-plugin-template": "^17.3.0",
-    "@angular-eslint/schematics": "^17.3.0",
-    "@angular-eslint/template-parser": "^17.3.0",
-    "@angular/compiler-cli": "^17.3.6",
-    "@angular/language-service": "^17.3.6",
+    "@angular-devkit/build-angular": "^18.0.2",
+    "@angular/compiler-cli": "^18.0.1",
+    "@angular/cli": "^18.0.2",
+    "@angular/language-service": "^18.0.1",
     "@commitlint/cli": "^16.0.1",
     "@commitlint/config-conventional": "^17",
     "@types/angular": "1.5.11",
@@ -116,12 +111,7 @@
     "autoprefixer": "~6",
     "canonical-path": "0.0.2",
     "concurrently": "^3.2.0",
-    "eslint": "^8.57.0",
-    "eslint-config-prettier": "^8.5.0",
-    "eslint-plugin-import": "2.26.0",
-    "eslint-plugin-jsdoc": "39.3.6",
-    "eslint-plugin-prefer-arrow": "1.2.3",
-    "eslint-plugin-prettier": "^5.0.1",
+    "eslint": "9",
     "grunt": "^1.0.4",
     "grunt-bump": "0.8.0",
     "grunt-coffeelint": "0.0.16",
@@ -159,7 +149,7 @@
     "sass": "^1.48.0",
     "tailwindcss": "~3.3",
     "ts-node": "~10.9",
-    "typescript": "~5.2",
+    "typescript": "~5.4.5",
     "underscore": "^1.8.3"
   },
   "optionalDependencies": {
