@@ -1,7 +1,7 @@
 import { TestBed, tick, fakeAsync } from '@angular/core/testing';
-import { HttpClientTestingModule, HttpTestingController } from '@angular/common/http/testing';
+import { HttpTestingController, provideHttpClientTesting } from '@angular/common/http/testing';
 import { User, UserService } from 'src/app/api/models/doubtfire-model';
-import { HttpRequest } from '@angular/common/http';
+import { HttpRequest, provideHttpClient, withInterceptorsFromDi } from '@angular/common/http';
 import { analyticsService } from 'src/app/ajs-upgraded-providers';
 
 describe('UserService', () => {
@@ -11,9 +11,9 @@ describe('UserService', () => {
 
   beforeEach(() => {
     TestBed.configureTestingModule({
-      imports: [HttpClientTestingModule],
-      providers: [UserService, { provide: analyticsService, useValue: analyticsServiceStub }],
-    });
+    imports: [],
+    providers: [UserService, { provide: analyticsService, useValue: analyticsServiceStub }, provideHttpClient(withInterceptorsFromDi()), provideHttpClientTesting()]
+});
 
     userService = TestBed.inject(UserService);
     httpMock = TestBed.inject(HttpTestingController);
