@@ -1,4 +1,4 @@
-import {HttpClient, HttpResponse} from '@angular/common/http';
+import { HttpClient, HttpResponse } from '@angular/common/http';
 import {Injectable} from '@angular/core';
 import {AlertService} from '../services/alert.service';
 
